migrate webhook IDs to 64-bit integers
Fixes #2443 (other fixes needed)

Related to the repo ID change made by Evan recently: use 64-bit ints for
webhook IDs. This involves increasing the size of integer we use in the
DB record, and making logic changes where needed.

+++ internal/controlplane/handlers_repositories.go
@@ -97,8 +97,8 @@ func (s *Server) RegisterRepository(ctx context.Context,
 		RepoID:    r.RepoId,
 		IsPrivate: r.IsPrivate,
 		IsFork:    r.IsFork,
-		WebhookID: sql.NullInt32{
-			Int32: int32(r.HookId),
+		WebhookID: sql.NullInt64{
+			Int64: r.HookId,
 			Valid: true,
 		},
 		CloneUrl:   r.CloneUrl,
